Fix initialisation bug for rmarkdown
`.projr_init_engine_rmd` was never run, so the RMarkdown document was never creatd.

This happened because when someone selected `RMarkdown document` as the
document engine, this saved the doc engine
as `rmd_document`.
However, later code assumed it would be `"rmd"` if it
was an RMarkdown project, so those were never activated.